Refactor test connection database and delete test files
Move test connection db to config folder for associate the  function (maintainability)

Now, focused on dev, the tests will be conduct later
